+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Add settings to disable egui's default navigation and activation.
+- Settings to disable egui's default navigation and activation.
+- `kbgp_focus_label` and `kbgp_set_focus_label` to control the focus.
 
 ## 0.2.1 - 2022-02-02
 ### Fixed